Implemented and tested naive perceptron

+++ src/main/scala/uk/co/matthewmcgowan/model/IndependentVoterVariable.scala
@@ -12,5 +12,5 @@ trait IndependentVoterVariable {
   val location: Location
   val income: Double
 
-  type Age = Int
+  type Age = Double
 }
